fixed form templating issues, moved forms to components
