+++ test/dummy/app/views/admin/people/new.html.erb
@@ -1,4 +1,4 @@
-<h2>Register a new person</h2>
+<div class="title-block"><h3 class="title">Register a new person</h3></div>
 <%= render partial: 'kadmin/helpers/form_errors', locals: { model: @person_form } %>
 <%= form_for [:admin, @person_form] do |form| %>
   <%= render partial: 'admin/people/form', locals: { form: form, model: @person_form } %>
